requirements.txt
ultralytics>=8.0.0
opencv-python
Pillow
python-dotenv
